refactor: 랜덤값 생성 기능 클래스 이동
- 랜덤값 생성기능은 Position 클래스에서 사용되는것이 더 적합하다고 판단하여 위치를 이동.
- 이로 인해 해당 기능의 테스트코드도 동시에 이동됨.
