+++ src/main/java/racingcar/Position.java
@@ -1,9 +1,13 @@
 package racingcar;
 
+import camp.nextstep.edu.missionutils.Randoms;
+
 public class Position {
     public int position;
     private static final int FORWARD_THRESHOLD=4;
     private static final String SLASH="-";
+    public static final int VALUE_LOW_BOUND=0;
+    public static final int VALUE_HIGH_BOUND=9;
 
     public Position(){
         position=0;
@@ -23,5 +27,9 @@ public class Position {
 
     public String switchPositionToSlash(){
         return SLASH.repeat(position);
+    }
+
+    public int makeRandomValue(){
+        return Randoms.pickNumberInRange(VALUE_LOW_BOUND,VALUE_HIGH_BOUND);
     }
 }
